[박소연] 과제 제출 - view 완성, jQuery 빈칸검사, ReportSubmit 유효성 검사 작성, Ajax 과제 제출 파일 업로드 Mapper, Controller, Service 작성 완료, application-db 추가, teamColor 변경, FileForm vo 추가, 첨부파일 Mapper 수정

+++ lms/src/main/java/com/gd/lms/vo/ReportSubmitFile.java
@@ -8,7 +8,7 @@ public class ReportSubmitFile {
 	private int reportSubmitNo;				// reportSubmit 번호
 	private String reportSubmitFilename;	// 제줄한 과제 첨부파일 새 이름
 	private String reportSubmitOriginName;	// 제줄한 과제 첨부파일 기존 이름
-	private int reportSubmitFileSize;		// 제줄한 과제 첨부파일 크기
+	private long reportSubmitFileSize;		// 제줄한 과제 첨부파일 크기
 	private String reportSubmitFileType; 	// 제줄한 과제 첨부파일 타입
 	private String createDate;				// 제줄한 과제 첨부파일 등록일
 }
